stubs' children count serializtion modified
git-svn-id: http://svn.jetbrains.org/idea/Trunk/scala@21798 59f64563-22e8-0310-8d2c-98a15e1539f2

+++ scala_core/src/org/jetbrains/plugins/scala/lang/psi/stubs/elements/ScExtendsBlockElementType.scala
@@ -17,7 +17,7 @@ class ScExtendsBlockElementType
 extends ScStubElementType[ScExtendsBlockStub, ScExtendsBlock]("extends block") {
 
   def serialize(stub: ScExtendsBlockStub, dataStream: StubOutputStream) {
-    dataStream.writeByte(stub.getBaseClasses.length)
+    dataStream.writeInt(stub.getBaseClasses.length)
     for (name <- stub.getBaseClasses) dataStream.writeName(name)
   }
 
@@ -28,7 +28,7 @@ extends ScStubElementType[ScExtendsBlockStub, ScExtendsBlock]("extends block") {
   }
 
   def deserializeImpl(dataStream: StubInputStream, parentStub: Any): ScExtendsBlockStub = {
-    val n = dataStream.readByte
+    val n = dataStream.readInt
     val baseClasses = new Array[String](n)
     for (i <- 0 to n-1) baseClasses(i) = StringRef.toString(dataStream.readName)
     new ScExtendsBlockStubImpl(parentStub.asInstanceOf[StubElement[PsiElement]], this, baseClasses)
